# openapi.yaml
openapi: 3.0.3
info:
  title: Sausalito API
  version: 1.0.0
  description: Sausalito 쇼핑몰 백엔드 API
paths:
...
  /api/v1/admin/bank-transfers:
    get:
      operationId: v1_admin_bank_transfers_retrieve
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT